fix(clash): widen clash element ids to 64-bit so a large id can't outlive its own delete (#732)
Revit 2024+ ElementId.Value is Int64 and a large or long-lived model mints
ids past int.MaxValue. ClashElementKey held them as int, so every such id was
truncated on entry to the clash tree — and the truncation is not merely lossy,
it collides: 2^32 + 7 and 7 both narrow to 7.

That leaked into federation. GeometrySyncHandler builds a ClashElementKey per
changed element, GlbSerializer stamped that key's id into the GLB node extras
as an int, and the server reads it with GetInt64 and keys FederatedElement on
it. A >int.MaxValue element was stored under a truncated id while the delete
that followed carried the real one (the delta path was widened in #684) — the
tombstone missed and stale geometry lingered in the federated scene.

Widened the whole clash-engine identity path in one pass, so nothing is left
half-narrow: ClashElementKey (fields, ctor, Equals, GetHashCode — now folding
the high word in rather than dropping it), ClashElementRecord, the element-id
maps and public surface of ClashSession, LiveClashUpdater's dirty queue (whose
deletion sentinel negates the id, so a narrowing cast could flip its meaning
too), LiveClashHandler, LiveClashFlag, ClashHistory's fuzzy bucket key,
ClashGrouper's pattern key, ClashNotifications, ClashSlaIntegration's owner
cache, and GlbSerializer's node extras. Dropped the (int)…Value casts at the
three points ids enter: ClashSession, ClashExportContext, GeometrySyncHandler.

No server-side change: FederatedElement.ElementId is already long, and the
server's own ClashRecord keys elements on IFC GUID strings, not numeric ids.

Tests: nine cases in StingTools.Clash.Tests/ClashElementId64BitTests.cs. The
load-bearing one serializes a real ClashMeshBuffer through GlbSerializer and
reads the node extras back with the same offsets and GetInt64 the server's
ParseGlbNodeExtras uses — so the plugin end of the wire is proven against the
reader that consumes it, not a hand-written GLB. The others pin the key's
identity and hash across a truncation collision, the clashes.json round-trip,
and the two places a collided id would silently merge two elements (history
fuzzy match, element-pattern grouping). GlbSerializer is now <Compile Include>d
into the clash test project; it imports Autodesk.Revit.DB but uses nothing from
it, like the four files RevitStubs already covers.

Closes #722.

+++ StingTools/Clash/ClashRunCommand.cs
@@ -383,7 +383,7 @@ namespace StingTools.Core.Clash
                 {
                     // Build the ElementId list once. Revit 2024+: ElementId(long).
                     var ids = new List<ElementId>(keys.Count);
-                    foreach (var k in keys) ids.Add(new ElementId((long)k.ElementId));
+                    foreach (var k in keys) ids.Add(new ElementId(k.ElementId));
                     // Single-collector pass scoped to these ids; constructs
                     // an Id → Element map without per-mesh API hits.
                     var byId = new Dictionary<long, Element>(keys.Count);
@@ -711,7 +711,7 @@ namespace StingTools.Core.Clash
         private static void WriteColdInitLiveFlags(Document doc, ClashRunRecord run)
         {
             if (doc == null || run?.Clashes == null) return;
-            var flagged = new HashSet<int>();
+            var flagged = new HashSet<long>();
             foreach (var c in run.Clashes)
             {
                 if (c.State == "Resolved" || c.State == "Void") continue;
@@ -724,7 +724,7 @@ namespace StingTools.Core.Clash
             // Empty cleared list — clearing is owned by the live session's
             // RefreshElement / RemoveElement and was already handled by
             // SeedFromRun's diff. We only need to assert flag=1 here.
-            LiveClashFlag.Apply(doc, flagged, Array.Empty<int>());
+            LiveClashFlag.Apply(doc, flagged, Array.Empty<long>());
             StingLog.Info($"WriteColdInitLiveFlags: wrote CLASH_LIVE_FLAG=1 on {flagged.Count} elements");
         }
